+++ src/main/java/pe/edu/ucsp/oms/repository/ContentDao.java
@@ -2,7 +2,6 @@ package pe.edu.ucsp.oms.repository;
 
 import pe.edu.ucsp.oms.domain.Content;
 
-
-public interface ContentDao extends GenericDao<Content, Integer>{
+public interface ContentDao extends GenericDao<Content, Long>{
 
 }
